src/microservices/apigw/requirements.txt
Django==1.11.10
djangorestframework==3.7.7
django-cors-headers==2.2.0
requests==2.18.4
PyJWT==1.5.3
